refactor(geometry-engine): apply machine-applicable clippy fixes
cargo clippy --fix pass across geometry-engine/src. Auto-fixes for:
- clone_on_copy (drop redundant .clone() on Copy types)
- useless_conversion (drop Vector3::from(Vector3) and Ok(_)? unwrap)
- unwrap_or_default (.or_insert_with(Vec::new) -> .or_default())
- new_without_default (add Default impls calling existing new())
- manual_swap (let temp; a=b; b=temp -> std::mem::swap)
- assign_op_pattern (a = a + b -> a += b)
- legacy_numeric_constants (${i,f}${ype}::MAX consts)
- unreadable_literal (1745329251994330 -> 1_745_329_251_994_33)
- needless_question_mark (Ok(x?) -> x)
- if/else where saturating_sub fits

Total: 2566 -> 2376 warnings (190 cleaned). 604/604 tests pass.
No semantic changes.

+++ roshera-backend/geometry-engine/src/operations/transform.rs
@@ -140,7 +140,7 @@ pub fn transform_faces(
     );
 
     Ok(TransformResult {
-        transformed_ids: faces.into_iter().map(|f| f as u32).collect(),
+        transformed_ids: faces.into_iter().map(|f| f).collect(),
         transform,
     })
 }
@@ -191,7 +191,7 @@ pub fn transform_edges(
     );
 
     Ok(TransformResult {
-        transformed_ids: edges.into_iter().map(|e| e as u32).collect(),
+        transformed_ids: edges.into_iter().map(|e| e).collect(),
         transform,
     })
 }
